added serviceImpl, and new method to service and serviceImpl
